pyproject.toml
[tool.poetry]
name = "reminder-bot"
version = "0.1.0"
description = ""
authors = ["Dennis M. <ezerex@mail.ru>"]
readme = "README.md"
packages = [{include = "reminder_bot"}]

[tool.poetry.dependencies]
python = "^3.11"
loguru = "^0.6.0"
pydantic = {extras = ["dotenv"], version = "^1.10.6"}
python-telegram-bot = {extras = ["ext"], version = "^20.1"}
ujson = "^5.7.0"
aiohttp = {extras = ["speedups"], version = "^3.8.4"}
fake-useragent = "^1.1.1"
sqlalchemy = "^2.0.6"
requests = "^2.28.2"
prettytable = "^3.6.0"


[tool.poetry.group.dev.dependencies]
ipython = "^8.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
